backend/requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.24,<1.0
pydantic>=2.6,<3.0
pydantic-settings>=2.2,<3.0
python-multipart>=0.0.9,<0.1
requests>=2.31,<3.0
pydantic[email]>=2.6,<3.0
dnspython>=2.4,<3.0
google-cloud-storage>=2.18.0
pandas>=2.2.2,<3.0
motor>=3.3,<4.0
openai>=1.43
